+++ zildo/src/main/java/zildo/monde/util/Zone.java
@@ -24,8 +24,8 @@ import zildo.monde.collision.Rectangle;
 
 public class Zone {
 
-	public int x1, y1;
-	public int x2, y2;
+	public float x1, y1;
+	public float x2, y2;
 
 	public Zone() {
 
@@ -48,7 +48,7 @@ public class Zone {
 	}
 
 	/** Build a zone with (x,y) (width, height) **/
-	public Zone(int x1, int y1, int x2, int y2) {
+	public Zone(float x1, float y1, float x2, float y2) {
 		this.x1 = x1;
 		this.x2 = x2;
 		this.y1 = y1;
